+++ src/constants.js
@@ -61,6 +61,15 @@ export const DEFAULT_COMPILER_ID = NO_PLUGIN_TYPE;
 
 export const DEFAULT_TESTER_ID = NO_PLUGIN_TYPE;
 
+export const DEFAULT_EXTENSIONS = {
+  'ext-docs-parser': {
+    config: {},
+    options: {
+      default: true
+    }
+  }
+};
+
 export const DEFAULT_DIST_DIRNAME = 'dist';
 
 export const DEFAULT_BUNDLE_FILENAME = 'dist.js';
@@ -77,7 +86,9 @@ export const NODE_PATH_SEPARATOR = '.';
 
 export const DEFAULT_DIR_STRUCTURE = `${BITS_DIRNAME}/{namespace}/{name}`;
 
-export const DEFAULT_DIR_DEPENDENCIES_STRUCTURE = `${BITS_DIRNAME}/.dependencies`;
+export const DEFAULT_DIR_DEPENDENCIES = '.dependencies';
+
+export const DEFAULT_DIR_DEPENDENCIES_STRUCTURE = `${BITS_DIRNAME}/${DEFAULT_DIR_DEPENDENCIES}`;
 
 export const DEFAULT_SEPARATOR = '/';
 
@@ -137,6 +148,60 @@ export const CFG_POST_IMPORT_HOOK_KEY = 'post_import_hook';
 export const CFG_CI_FUNCTION_PATH_KEY = 'ci_function_path';
 
 export const CFG_CI_ENABLE_KEY = 'ci_enable';
+
+/**
+ * bit hooks
+ */
+export const PRE_TAG_HOOK = 'pre-tag';
+
+export const POST_TAG_HOOK = 'post-tag';
+
+export const PRE_TAG_ALL_HOOK = 'pre-tag-all';
+
+export const POST_TAG_ALL_HOOK = 'post-tag-all';
+
+export const PRE_IMPORT_HOOK = 'pre-import';
+
+export const POST_IMPORT_HOOK = 'post-import';
+
+export const PRE_EXPORT_HOOK = 'pre-export';
+
+export const POST_EXPORT_HOOK = 'post-export';
+
+export const PRE_SEND_OBJECTS = 'pre-send-objects'; // pre-fetch
+
+export const POST_SEND_OBJECTS = 'post-send-objects'; // post-fetch
+
+export const PRE_RECEIVE_OBJECTS = 'pre-receive-objects'; // pre-put
+
+export const POST_RECEIVE_OBJECTS = 'post-receive-objects'; // post-put
+
+export const PRE_DEPRECATE_REMOTE = 'pre-deprecate-remote';
+
+export const POST_DEPRECATE_REMOTE = 'post-deprecate-remote';
+
+export const PRE_REMOVE_REMOTE = 'pre-remove-remote';
+
+export const POST_REMOVE_REMOTE = 'post-remove-remote';
+
+export const HOOKS_NAMES = [
+  PRE_TAG_HOOK,
+  POST_TAG_HOOK,
+  PRE_TAG_ALL_HOOK,
+  POST_TAG_ALL_HOOK,
+  PRE_IMPORT_HOOK,
+  POST_IMPORT_HOOK,
+  PRE_EXPORT_HOOK,
+  POST_EXPORT_HOOK,
+  PRE_SEND_OBJECTS,
+  POST_SEND_OBJECTS,
+  PRE_RECEIVE_OBJECTS,
+  POST_RECEIVE_OBJECTS,
+  PRE_DEPRECATE_REMOTE,
+  POST_DEPRECATE_REMOTE,
+  PRE_REMOVE_REMOTE,
+  POST_REMOVE_REMOTE
+];
 
 /**
  * cache root directory