+++ src/consumer/bit-json/abstract-bit-json.js
@@ -10,7 +10,8 @@ import {
   DEFAULT_DEPENDENCIES,
   NO_PLUGIN_TYPE,
   DEFAULT_LANGUAGE,
-  DEFAULT_BINDINGS_PREFIX
+  DEFAULT_BINDINGS_PREFIX,
+  DEFAULT_EXTENSIONS
 } from '../../constants';
 
 export type AbstractBitJsonProps = {
@@ -20,7 +21,8 @@ export type AbstractBitJsonProps = {
   tester?: string,
   dependencies?: Object,
   lang?: string,
-  bindingPrefix?: string
+  bindingPrefix?: string,
+  extensions?: Object
 };
 
 export default class AbstractBitJson {
@@ -33,8 +35,9 @@ export default class AbstractBitJson {
   dependencies: { [string]: string };
   lang: string;
   bindingPrefix: string;
+  extensions: Object;
 
-  constructor({ impl, spec, compiler, tester, dependencies, lang, bindingPrefix }: AbstractBitJsonProps) {
+  constructor({ impl, spec, compiler, tester, dependencies, lang, bindingPrefix, extensions }: AbstractBitJsonProps) {
     this.impl = impl || DEFAULT_IMPL_NAME;
     this.spec = spec || DEFAULT_SPECS_NAME;
     this.compiler = compiler || DEFAULT_COMPILER_ID;
@@ -42,6 +45,7 @@ export default class AbstractBitJson {
     this.dependencies = dependencies || DEFAULT_DEPENDENCIES;
     this.lang = lang || DEFAULT_LANGUAGE;
     this.bindingPrefix = bindingPrefix || DEFAULT_BINDINGS_PREFIX;
+    this.extensions = extensions || DEFAULT_EXTENSIONS;
   }
 
   get compilerId(): string {
@@ -117,7 +121,8 @@ export default class AbstractBitJson {
           compiler: this.compilerId,
           tester: this.testerId
         },
-        dependencies: this.dependencies
+        dependencies: this.dependencies,
+        extensions: this.extensions
       },
       isPropDefaultOrNull
     );